+++ app/src/main/java/com/shapeshed/aerial/ui/theme/Theme.kt
@@ -3,6 +3,7 @@ package com.shapeshed.aerial.ui.theme
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Typography
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.dynamicDarkColorScheme
 import androidx.compose.material3.dynamicLightColorScheme
@@ -79,7 +80,7 @@ fun AerialTheme(
 
     MaterialTheme(
         colorScheme = colorScheme,
-        typography = AerialTypography,
+        typography = Typography(),
         content = content,
     )
 }